requirements.txt
discord.py
yt-dlp
python-dotenv
requests
lyricsgenius
beautifulsoup4